+++ apps/client/src/components/aas/PropertyEditor.vue
@@ -38,7 +38,7 @@ const { can } = useAasAbility({
   getAccessPermissionRules: props.getAccessPermissionRules,
 });
 const disableEdit = computed(() => {
-  return !can(Permissions.Edit, props.path.idShortPathIncludingSubmodel ?? "");
+  return props.isArchived || !can(Permissions.Edit, props.path.idShortPathIncludingSubmodel ?? "");
 });
 
 async function submit() {