+++ apps/client/src/components/aas/SubmodelEditor.vue
@@ -46,7 +46,7 @@ const { can } = useAasAbility({
 });
 
 const disableEdit = computed(() => {
-  return !can(Permissions.Edit, props.path.idShortPathIncludingSubmodel ?? "");
+  return props.isArchived || !can(Permissions.Edit, props.path.idShortPathIncludingSubmodel ?? "");
 });
 
 defineExpose<{