+++ apps/main/src/aas/infrastructure/serialization/export-schemas/aas-export-types.ts
@@ -1,13 +1,15 @@
 import { z } from "zod";
 import { aasExportSchemaJsonV1_0 } from "./aas-export-v1.schema";
 import { aasExportSchemaJsonV2_0 } from "./aas-export-v2.schema";
+import { aasExportSchemaJsonV3_0 } from "./aas-export-v3.schema";
 
 export const AasExportSchemas = z.discriminatedUnion("version", [
   aasExportSchemaJsonV1_0,
   aasExportSchemaJsonV2_0,
+  aasExportSchemaJsonV3_0,
 ]);
 
 export type AasExport = z.infer<typeof AasExportSchemas>;
 
-export const aasExportSchemaJsonLatest = aasExportSchemaJsonV2_0;
+export const aasExportSchemaJsonLatest = aasExportSchemaJsonV3_0;
 export type AasExportLatestVersion = z.infer<typeof aasExportSchemaJsonLatest>;