+++ apps/main/src/open-api-docs/index.ts
@@ -2,6 +2,7 @@ import type { INestApplication } from "@nestjs/common";
 import { OpenAPIObject, SwaggerModule } from "@nestjs/swagger";
 import { createDocument } from "zod-openapi";
 import { aasPaths } from "./aas.paths";
+import { brandingPaths } from "./branding.path";
 
 const document = createDocument({
   openapi: "3.1.0",
@@ -17,6 +18,7 @@ const document = createDocument({
   ],
   paths: {
     ...aasPaths,
+    ...brandingPaths,
   },
   components: {
     parameters: {